+++ src/NClient/NClient.AspNetProxy.Standalone/ControllerClientProvider.cs
@@ -5,7 +5,7 @@ using Microsoft.Extensions.Logging;
 using NClient.Abstractions.Clients;
 using NClient.Abstractions.HttpClients;
 using NClient.Abstractions.Resilience;
-using NClient.AspNetProxy.Attributes;
+using NClient.AspNetProxy.Mappers;
 using NClient.AspNetProxy.Validators;
 using NClient.Core.Helpers;
 using NClient.Core.Interceptors;
