+++ src/NClient/NClient.Core/Exceptions/Factories/OuterExceptionFactory.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Linq;
 using System.Net;
 using System.Reflection;
